refactor(infrastructure): move range observable collection
