+++ CefFlashBrowser/ViewModels/SaveMgrWorkspaceViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using CefFlashBrowser.Models;
+using CefFlashBrowser.Infrastructure.Collections;
 using CefFlashBrowser.Utils;
 using SimpleMvvm;
 using SimpleMvvm.Command;
